- Adds holographicdisplay as a softdepends.   - If enabled in the config, and HolographicDisplays is present then rather than a sign on top of a deathchest, a hologram with the same information will be spawned.   - Holograms are stored in a new holograms.db which is saved to on shutdown as well as after a tombblock is made.   - Holograms are loaded at startup. - Moved some more stuff out of the EntityListener and into CenotaphUtil. - Moved all plugin-handling classes into a PluginHandlers package.
