+++ pom.xml
@@ -30,6 +30,10 @@
             <id>dynamap</id>
             <url>http://repo.mikeprimm.com</url>
         </repository>
+        <repository>
+		    <id>codemc-repo</id>
+		    <url>https://repo.codemc.org/repository/maven-public/</url>
+		</repository>
     </repositories>
 
 
@@ -57,7 +61,6 @@
             <version>7.0.0-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
-
         <dependency>
             <groupId>com.sk89q.worldedit</groupId>
             <artifactId>worldedit-bukkit</artifactId>
@@ -70,6 +73,12 @@
             <artifactId>jansi</artifactId>
             <version>1.18</version>
         </dependency>
+        <dependency>
+		    <groupId>com.gmail.filoghost.holographicdisplays</groupId>
+		    <artifactId>holographicdisplays-api</artifactId>
+		    <version>2.3.2</version>
+		    <scope>provided</scope>
+		</dependency>
     </dependencies>
 
     <build>
