+++ src/main/java/com/MoofIT/Minecraft/Cenotaph/PluginHandlers/WorldGuardWrapper.java
@@ -1,4 +1,4 @@
-package com.MoofIT.Minecraft.Cenotaph;
+package com.MoofIT.Minecraft.Cenotaph.PluginHandlers;
 
 /**
  * Cenotaph
@@ -24,6 +24,7 @@ import com.sk89q.worldguard.internal.platform.WorldGuardPlatform;
 import com.sk89q.worldguard.protection.flags.Flags;
 import org.bukkit.entity.Player;
 
+import com.MoofIT.Minecraft.Cenotaph.Cenotaph;
 import com.sk89q.worldguard.LocalPlayer;
 import com.sk89q.worldguard.bukkit.WorldGuardPlugin;
 import com.sk89q.worldguard.protection.ApplicableRegionSet;
